Extract ablation logic to modeling/ablation.py, move CLI scripts to scripts/
run_ablation_matrix.py and run_top_features_eval.py are now thin wrappers
in scripts/. Shared helpers and run_ablation_matrix() live in modeling/ablation.py,
which is where the modeling/ array workers (run_arm_fold, build_cached_table,
merge_results) already needed them.
